+++ synapseclient/core/version_check.py
@@ -11,8 +11,8 @@ Print release notes for installed version of client::
 
     synapseclient.release_notes()
 
-.. automethod:: synapseclient.version_check.check_for_updates
-.. automethod:: synapseclient.version_check.release_notes
+.. automethod:: synapseclient.core.version_check.check_for_updates
+.. automethod:: synapseclient.core.version_check.release_notes
 
 """
 
@@ -107,8 +107,10 @@ def release_notes(version_url=None):
     is supplied.
 
     :param version_url: Defaults to None, meaning release notes for the installed version. Alternatives are:
+
                             - synapseclient.version_check._VERSION_URL
                             - synapseclient.version_check._DEV_VERSION_URL
+
     """
     version_info = _get_version_info(version_url)
     sys.stderr.write('Python Synapse Client version %s release notes\n\n' % version_info['latestVersion'])